services/recommend/requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]

# Database (PostgreSQL)
psycopg2-binary

# Configuration & Environment
python-dotenv

# Data Validation
pydantic